+++ plugin-server/src/cdp/services/hogflows/hogflow-executor.service.ts
@@ -1,9 +1,9 @@
 import { DateTime } from 'luxon'
 
-import { HogFlow } from '../../schema/hogflow'
-import { Hub } from '../../types'
-import { logger } from '../../utils/logger'
-import { UUIDT } from '../../utils/utils'
+import { HogFlow } from '../../../schema/hogflow'
+import { Hub } from '../../../types'
+import { logger } from '../../../utils/logger'
+import { UUIDT } from '../../../utils/utils'
 import {
     CyclotronJobInvocationHogFlow,
     CyclotronJobInvocationResult,
@@ -11,11 +11,11 @@ import {
     HogFunctionInvocationGlobals,
     LogEntry,
     MinimalAppMetric,
-} from '../types'
-import { convertToHogFunctionFilterGlobal } from '../utils'
-import { filterFunctionInstrumented } from '../utils/hog-function-filtering'
-import { createInvocationResult } from '../utils/invocation-utils'
-import { HogFlowActionRunner } from './hogflows/actions'
+} from '../../types'
+import { convertToHogFunctionFilterGlobal } from '../../utils'
+import { filterFunctionInstrumented } from '../../utils/hog-function-filtering'
+import { createInvocationResult } from '../../utils/invocation-utils'
+import { HogFlowActionRunner } from './actions'
 
 export const MAX_ACTION_STEPS_HARD_LIMIT = 1000
